aster_systree

Macro inherit_sys_leaf_node

macro_rules! inherit_sys_leaf_node {
    ($struct_name:ident, $field:ident, {$($fn_override:item)*}) => { ... };
}

A convenience macro for implementing SysObj and SysNode trait for a SysTree leaf type through inheriting the implementation of the target field.

Users must ensures the target field is a NormalNodeFields type.

This macro takes three parameters. The first two parameters correspond to:

The name of the struct that implements the SysObj trait,

The name of the field to be inherited,

The third parameter is used to implement the required trait methods or override default implementations. For SysObj and SysNode traits, some of their methods require users to implement them within this macro, some methods provide default implementations that users can override, and some methods provide default implementations that users cannot override. The breakdown is as follows:

Methods that require users to implement:

SysNode::perms

Methods with default implementations that users can override:

SysObj::is_root

SysObj::init_parent

SysNode::is_attr_absent

SysNode::read_attr

SysNode::write_attr

SysNode::read_attr_at

SysNode::write_attr_at

Other methods are provided with default implementations that users cannot override.

Here, “default implementations” refer to either directly inheriting the implementation of a method with the same name from the target field or, in the absence of a method with the same name, the default implementation provided by the trait.

Note that for the SysNode trait, read_attr can be automatically implemented in terms of read_attr_at, and write_attr_at can be automatically implemented in terms of write_attr since most sysfs attributes do not support partial writes and will ignore the offset. Therefore, it is recommended to override the read_attr_at and write_attr methods. In addition, users can use [aster_util::printer::VmPrinter] to easily handle the offset when overriding the read_attr_at method.

Examples

/// // A struct representing a `SysTree` leaf node.
struct LeafNode {
    fields: NormalNodeFields<dyn SysObj, Self>,
    // ... other fields
}

inherit_sys_leaf_node!(
  LeafNode,      // The struct name.
  fields,        // The field to be inherited.
  {
     fn perms(&self) -> SysPerms {
       SysPerms::DEFAULT_RW_PERMS
     }
  } // Override the `perms` method.
);