aster_systree

Trait SysBranchNode

pub trait SysBranchNode: SysNode {
    // Required methods
    fn visit_child_with(
        &self,
        name: &str,
        f: &mut dyn FnMut(Option<&Arc<dyn SysObj>>),
    );
    fn visit_children_with(
        &self,
        min_id: u64,
        f: &mut dyn for<'a> FnMut(&'a Arc<dyn SysObj>) -> Option<()>,
    );
    fn child(&self, name: &str) -> Option<Arc<dyn SysObj>>;

    // Provided methods
    fn children(&self) -> Vec<Arc<dyn SysObj>> { ... }
    fn count_children(&self) -> usize { ... }
    fn create_child(&self, _name: &str) -> Result<Arc<dyn SysObj>> { ... }
    fn remove_child(&self, _name: &str) -> Result<Arc<dyn SysObj>> { ... }
}

A trait that represents a branching node in a SysTree.

Required Methods

fn visit_child_with( &self, name: &str, f: &mut dyn FnMut(Option<&Arc<dyn SysObj>>), )

Visits a child node with the given name using a closure.

If the child with the given name exists, a reference to the child will be provided to the closure. Otherwise, the closure will be given a None.

Efficiency

This method is a more efficient, but less convenient version of the child method. The method does not require taking the ownership of the child node. So use this method when efficiency is a primary concern, while using the child method for the sake of convenience.

Deadlock

The implementation of this method depends on the concrete type and probably will hold an internal lock. So the caller should do as little as possible inside the closure. In particular, the caller should not invoke other methods on this object as this might cause deadlock.

fn visit_children_with( &self, min_id: u64, f: &mut dyn for<'a> FnMut(&'a Arc<dyn SysObj>) -> Option<()>, )

Visits child nodes with a minimum ID using a closure.

This method iterates over the child nodes whose IDs are no less than a specified minimum value. and provide them to the given closure one at a time.

The iteration terminates until there are no unvisited children or the closure returns a None.

Efficiency

This method is a more efficient, but less convenient version of the children method. The method require neither taking the ownership of the child nodes nor doing heap allocations. So use this method when efficiency is a primary concern, while using the children method for the sake of convenience.

Deadlock

Same as the visit_child_with method.

fn child(&self, name: &str) -> Option<Arc<dyn SysObj>>

Returns a child with a specified name.

Provided Methods

fn children(&self) -> Vec<Arc<dyn SysObj>>

Collects all children into a Vec.

fn count_children(&self) -> usize

Counts the number of children.

fn create_child(&self, _name: &str) -> Result<Arc<dyn SysObj>>

Creates a new child node with the given name.

This new child will be added to this branch node.

fn remove_child(&self, _name: &str) -> Result<Arc<dyn SysObj>>

Removes a child node with the given name.