aster_systree

Trait SysNode

pub trait SysNode: SysObj {
    // Required methods
    fn node_attrs(&self) -> &SysAttrSet;
    fn is_attr_absent(&self, _name: &str) -> bool;
    fn read_attr(&self, name: &str, writer: &mut VmWriter<'_>) -> Result<usize>;
    fn write_attr(&self, name: &str, reader: &mut VmReader<'_>) -> Result<usize>;
    fn read_attr_at(
        &self,
        name: &str,
        offset: usize,
        writer: &mut VmWriter<'_>,
    ) -> Result<usize>;
    fn write_attr_at(
        &self,
        name: &str,
        _offset: usize,
        reader: &mut VmReader<'_>,
    ) -> Result<usize>;
    fn perms(&self) -> SysPerms;

    // Provided methods
    fn show_attr(&self, name: &str) -> Result<String> { ... }
    fn store_attr(&self, name: &str, new_val: &str) -> Result<usize> { ... }
}

The trait that abstracts a “normal” node in a SysTree.

The branching and leaf nodes are considered “normal”, whereas the symlink nodes are considered “special”. This trait abstracts the common interface of “normal” nodes. In particular, every “normal” node may have associated attributes.

Required Methods

fn node_attrs(&self) -> &SysAttrSet

Returns the attribute set of a SysNode.

The attribute set returned by this method contains all possible attributes that this node may have. For nodes with dynamically changing attributes, it is necessary to additionally use the SysNode::is_attr_absent method to confirm whether an attribute is currently present.

fn is_attr_absent(&self, _name: &str) -> bool

Returns whether an attribute with the given name is absent.

If the node does not have the attribute with the given name, the return value is unspecified.

fn read_attr(&self, name: &str, writer: &mut VmWriter<'_>) -> Result<usize>

Reads the value of an attribute.

fn write_attr(&self, name: &str, reader: &mut VmReader<'_>) -> Result<usize>

Writes the value of an attribute.

fn read_attr_at( &self, name: &str, offset: usize, writer: &mut VmWriter<'_>, ) -> Result<usize>

Reads the value of an attribute from the specified offset.

fn write_attr_at( &self, name: &str, _offset: usize, reader: &mut VmReader<'_>, ) -> Result<usize>

Writes the value of an attribute at the specified offset.

fn perms(&self) -> SysPerms

Returns the initial permissions of a node.

The FS layer should take the value returned from this method as this initial permissions for the corresponding inode.

Provided Methods

fn show_attr(&self, name: &str) -> Result<String>

Shows the string value of an attribute.

Most attributes are textual, rather binary. So using this show_attr method is more convenient than the read_attr method.

fn store_attr(&self, name: &str, new_val: &str) -> Result<usize>

Stores the string value of an attribute.

Most attributes are textual, rather binary. So using this store_attr method is more convenient than the write_attr method.