aster_systree

Trait SysObj

pub trait SysObj:
    Any
    + Send
    + Sync
    + Debug
    + 'static {
    // Required methods
    fn as_any(&self) -> &dyn Any;
    fn id(&self) -> &SysNodeId;
    fn type_(&self) -> SysNodeType;
    fn name(&self) -> &SysStr;
    fn init_parent(&self, parent: Weak<dyn SysBranchNode>);
    fn parent(&self) -> Option<Arc<dyn SysBranchNode>>;

    // Provided methods
    fn cast_to_symlink(&self) -> Option<Arc<dyn SysSymlink>> { ... }
    fn cast_to_node(&self) -> Option<Arc<dyn SysNode>> { ... }
    fn cast_to_branch(&self) -> Option<Arc<dyn SysBranchNode>> { ... }
    fn is_root(&self) -> bool { ... }
    fn path(&self) -> SysStr { ... }
}

The base trait for any node in a SysTree.

Required Methods

fn as_any(&self) -> &dyn Any

Returns a reference to this object as Any for downcasting.

fn id(&self) -> &SysNodeId

Returns the unique and immutable ID of a node.

fn type_(&self) -> SysNodeType

Returns the type of a node.

fn name(&self) -> &SysStr

Returns the name of a node.

The name is guaranteed not to contain two special characters: '/' and '\0'.

The root node of a SysTree has an empty name. All other inodes must have an non-empty name.

fn init_parent(&self, parent: Weak<dyn SysBranchNode>)

Initializes the parent in the SysTree.

An appropriate timing to call this method is when a SysTree node is added to another node as a child.

Panics

This method should be called at most once; otherwise, it may trigger panicking.

fn parent(&self) -> Option<Arc<dyn SysBranchNode>>

Returns the parent in the SysTree.

Returns None if the parent of the node has not been initialized yet.

Provided Methods

fn cast_to_symlink(&self) -> Option<Arc<dyn SysSymlink>>

Casts this object to a trait object of SysTree symlink.

fn cast_to_node(&self) -> Option<Arc<dyn SysNode>>

Casts this object to a trait object of a SysTree node.

fn cast_to_branch(&self) -> Option<Arc<dyn SysBranchNode>>

Casts this object to a trait object of a SysTree branch node.

fn is_root(&self) -> bool

Returns whether a node is the root of a SysTree.

fn path(&self) -> SysStr

Returns the path.

The path of a SysTree node is defined as below:

If a node is the root, then the path is /.

If a node is a non-root, we will need to see if it has a parent:

If the node has a parent, then the path is defined as the parent path and the node’s name joined with a / in between;

Otherwise, the node’s name is taken as its path.